Added a QR Code to the online version of this resume.
Adding a QR Code that contains a link to the hosted github PDF file will
allow people to get the latest version if need be.

A rename of the LaTeX and PDF template was also performed, to have a more
intuitive name.

Added a few Git ignore entries as well to mask all generated but
unnecessary LaTeX files.